fixed error: ^C would exit the shell when used on the first prompt
