+++ inc/minishell.h
@@ -6,25 +6,30 @@
 /*   By: mortins- <mortins-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/06/05 16:01:34 by ddiniz-m          #+#    #+#             */
-/*   Updated: 2023/06/13 18:27:10 by mortins-         ###   ########.fr       */
+/*   Updated: 2023/06/16 17:00:20 by mortins-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef MINISHELL_H
 # define MINISHELL_H
 
-//--------------------------------INCLUDES-------------------------------------
-#include "../libft/src/libft.h"
+//----------------------------------INCLUDES-----------------------------------
+# include "../libft/src/libft.h"
 
-#include <stdio.h>
-#include <stdlib.h>
-#include <unistd.h>
+# include <stdio.h>
+# include <stdlib.h>
+# include <unistd.h>
 
 //readline, rl_on_new_line, rl_replace_line, rl_redisplay
-#include <readline/readline.h>
-#include <readline/history.h>
+# include <readline/readline.h>
+# include <readline/history.h>
 
 //signal
-#include <signal.h>
+# include <signal.h>
+
+//-----------------------------------SIGNALS-----------------------------------
+void	signal_init(void);
+void	signal_interrupt(int signum);
+void	signal_exit(char *str);
 
 #endif
