+++ src/signals.c
@@ -1,0 +1,48 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   signals.c                                          :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: mortins- <mortins-@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2023/06/16 16:08:59 by mortins-          #+#    #+#             */
+/*   Updated: 2023/06/16 16:59:07 by mortins-         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "../inc/minishell.h"
+
+//	CTRL-C sends an "Interrupt" signal.
+//	CTRL-\ sends a "Quit" signal.
+//	When using SIG_IGN in signal(), it ignores the signal received.
+void	signal_init(void)
+{
+	signal(SIGINT, signal_interrupt);
+	signal(SIGQUIT, SIG_IGN);
+}
+
+//	CTRL-C sends an "Interrupt" signal.
+//	For documentation on rl_functions, read:
+//	https://tiswww.case.edu/php/chet/readline/readline.html#Command-Line-Editing
+void	signal_interrupt(int signum)
+{
+	if (signum == SIGINT)
+	{
+		printf("\n");
+		rl_on_new_line();
+		rl_replace_line("", 0);
+		rl_redisplay();
+	}	
+}
+
+//	CTRL-D represents "No input".
+//	When pressed, str is NULL and it exits the shell.
+void	signal_exit(char *str)
+{
+	if (!str)
+	{
+		printf("exit\n");
+		free(str);
+		exit (0);
+	}
+}
